refactor filter definitions into separate typescript files
