+++ src/store/modules/queryFilters/filterDefinitions/filterDefinitionInterface.ts
@@ -1,0 +1,11 @@
+export interface filterDefinition {
+  label: string;
+  slug: string;
+  apiKey: string;
+  items: {
+    title: string;
+    value?: string;
+    default?: boolean;
+    apiValue: string | string[];
+  }[];
+}
