+++ src/store/modules/queryFilters/filterDefinitions/index.ts
@@ -1,0 +1,19 @@
+import { typeFilterDefinition } from "./typeFilter";
+import { documentsTypeFilterDefinition } from "./documentsTypeFilter";
+import { imagesTypeFilterDefinition } from "./imagesTypeFilter";
+import { audioTypeFilterDefinition } from "./audioTypeFilter";
+import { videoTypeFilterDefinition } from "./videoTypeFilter";
+import { otherTypeFilterDefinition } from "./otherTypeFilter";
+import { languageFilterDefinition } from "./languageFilter";
+import { sizeFilterDefinition } from "./sizeFilter";
+
+export default {
+  typeFilterDefinition,
+  documentsTypeFilterDefinition,
+  imagesTypeFilterDefinition,
+  audioTypeFilterDefinition,
+  videoTypeFilterDefinition,
+  otherTypeFilterDefinition,
+  languageFilterDefinition,
+  sizeFilterDefinition,
+};
